+++ taskrc/ast/taskrc.go
@@ -24,6 +24,7 @@ type Remote struct {
 	Offline      *bool          `yaml:"offline"`
 	Timeout      *time.Duration `yaml:"timeout"`
 	CacheExpiry  *time.Duration `yaml:"cache-expiry"`
+	CacheDir     *string        `yaml:"cache-dir"`
 	TrustedHosts []string       `yaml:"trusted-hosts"`
 }
 
@@ -46,7 +47,7 @@ func (t *TaskRC) Merge(other *TaskRC) {
 	t.Remote.Offline = cmp.Or(other.Remote.Offline, t.Remote.Offline)
 	t.Remote.Timeout = cmp.Or(other.Remote.Timeout, t.Remote.Timeout)
 	t.Remote.CacheExpiry = cmp.Or(other.Remote.CacheExpiry, t.Remote.CacheExpiry)
-
+	t.Remote.CacheDir = cmp.Or(other.Remote.CacheDir, t.Remote.CacheDir)
 	if len(other.Remote.TrustedHosts) > 0 {
 		merged := slices.Concat(other.Remote.TrustedHosts, t.Remote.TrustedHosts)
 		slices.Sort(merged)